remove unused

+++ editor/UserInterface/TimelineSlider.cs
@@ -156,7 +156,7 @@ public class TimelineSlider : Slider
 
             if (snap != 0 || (tickCount == 0 && timingPoint.OmitFirstBarLine) || beatCount % timingPoint.BeatPerMeasure != 0) lineSize.Y *= .5f;
 
-            var tickX = offset.X + (float)Manager.SnapToPixel((time - leftTime) * timeScale);
+            var tickX = offset.X + Manager.SnapToPixel((float)(time - leftTime) * timeScale);
             var tickOpacity = tickX > beatmapLabel.TextBounds.Left - 8 ? actualOpacity * .2f : actualOpacity;
 
             drawLine(drawContext, new(tickX, offset.Y + lineBottomY), lineSize, tickColor, tickOpacity);
@@ -165,12 +165,12 @@ public class TimelineSlider : Slider
         // HitObjects
         if (project.ShowHitObjects) foreach (var hitObject in project.MainBeatmap.HitObjects) if (leftTime < hitObject.EndTime && hitObject.StartTime < rightTime)
                 {
-                    var left = Math.Max(0, (hitObject.StartTime - leftTime) * timeScale);
-                    var right = Math.Min((hitObject.EndTime - leftTime) * timeScale, Bounds.Width);
+                    var left = Math.Max(0, (float)(hitObject.StartTime - leftTime) * timeScale);
+                    var right = Math.Min((float)(hitObject.EndTime - leftTime) * timeScale, Bounds.Width);
                     var height = Math.Max(Bounds.Height * .1f - pixelSize, pixelSize);
 
-                    drawLine(drawContext, offset + new Vector2((float)Manager.SnapToPixel(left - height / 2), hitObjectsY),
-                        new((float)Manager.SnapToPixel(right - left + height), height), hitObject.Color, actualOpacity);
+                    drawLine(drawContext, offset + new Vector2(Manager.SnapToPixel(left - height / 2), hitObjectsY),
+                        new(Manager.SnapToPixel(right - left + height), height), hitObject.Color, actualOpacity);
                 }
 
         // Bookmarks
@@ -179,7 +179,7 @@ public class TimelineSlider : Slider
             drawLine(drawContext, new(timeToXTop(bookmark * .001f), offset.Y + Bounds.Height * .1f), new(pixelSize, Bounds.Height * .3f), bookmarkColor, actualOpacity);
 
             if (leftTime < bookmark && bookmark < rightTime) drawLine(drawContext,
-                offset + new Vector2((float)Manager.SnapToPixel((bookmark - leftTime) * timeScale), lineBottomY),
+                offset + new Vector2(Manager.SnapToPixel((bookmark - leftTime) * timeScale), lineBottomY),
                 new(pixelSize, Bounds.Height * .5f), bookmarkColor, actualOpacity);
         }
 
@@ -208,8 +208,8 @@ public class TimelineSlider : Slider
     }
     float timeToXTop(double time)
     {
-        var progress = (time - MinValue) / (MaxValue - MinValue);
-        return (float)Manager.SnapToPixel(AbsolutePosition.X + progress * Width);
+        var progress = (float)(time - MinValue) / (MaxValue - MinValue);
+        return Manager.SnapToPixel(AbsolutePosition.X + progress * Width);
     }
     void drawLine(DrawContext drawContext, Vector2 position, Vector2 size, Color color, float opacity)
     {
